Make userpreference using player data class and insert user data to database

+++ app/src/main/java/com/catnip/rockpaperscissorchapter6and7/ui/game/mode/dialog/PlayerMenusDialogFragment.kt
@@ -18,7 +18,7 @@ class PlayerMenusDialogFragment :
         FragmentPlayerMenusBinding::inflate
     ), PlayerMenusContract.View {
 
-    private var players: MutableMap<Int,String> = mutableMapOf()
+    private var players: MutableMap<Long,String> = mutableMapOf()
     private var playerNames : MutableList<String> = mutableListOf()
 
     override fun onResume() {
